Introduce Domain module
