+++ lib/eventory.rb
@@ -13,8 +13,7 @@ require 'eventory/event_builder'
 require 'eventory/event_stream_processor'
 require 'eventory/projector'
 require 'eventory/reactor'
-require 'eventory/aggregate_root'
-require 'eventory/aggregate_repository'
+require 'eventory/domain'
 
 module Eventory
 end
