+++ lib/eventory/domain.rb
@@ -1,0 +1,7 @@
+require 'eventory/domain/aggregate_repository'
+require 'eventory/domain/aggregate_root'
+
+module Eventory
+  module Domain
+  end
+end
